aider: Ensure test directories are created before creating the 1KB test binary file.
# Aider chat conversation:

ASSISTANT: The error message indicates that the file `./tests/testfiles/1kb.bin` does not exist. This file is created in the `tests/test_functional_tests.py` script and is used for testing binary file operations.

The file is created in this block of code:

```python
# Create a 1KB test binary file
""" ./tests/testfiles/1kb.bin - 1KB binary file to be used in binary file tests """
with open("./tests/testfiles/1kb.bin", "wb") as f:
    f.write(os.urandom(1024))
```

The error could be due to the `./tests/testfiles` directory not existing when the script tries to create the `1kb.bin` file. The script does include a function `test_create_test_dirs()` to create this directory, but it might not be getting called before the `1kb.bin` file is created.

To fix this, we can ensure that the `./tests/testfiles` directory is created before the `1kb.bin` file. We can do this by calling the `test_create_test_dirs()` function at the start of the script.

Here is the *SEARCH/REPLACE block* for this change:

```python
tests/test_functional_tests.py
<<<<<<< SEARCH
# Create a 1KB test binary file
""" ./tests/testfiles/1kb.bin - 1KB binary file to be used in binary file tests """
with open("./tests/testfiles/1kb.bin", "wb") as f:
    f.write(os.urandom(1024))
=======
# Ensure test directories are created
test_create_test_dirs()

# Create a 1KB test binary file
""" ./tests/testfiles/1kb.bin - 1KB binary file to be used in binary file tests """
with open("./tests/testfiles/1kb.bin", "wb") as f:
    f.write(os.urandom(1024))
>>>>>>> REPLACE
```

This change will ensure that the `./tests/testfiles` directory is created before the script tries to create the `1kb.bin` file, which should prevent the `FileNotFoundError`.